+++ charbot/giveaway/__init__.py
@@ -1,0 +1,5 @@
+__all__ = ("Giveaway", "BidModal", "GiveawayView")
+
+from .cog import Giveaway
+from .modal import BidModal
+from .view import GiveawayView
